+++ client/src/lib/json-to-iac/power-vs-instances.js
@@ -101,6 +101,10 @@ function powerVsInstanceData(instance) {
   }
   delete data.sap;
   delete data.sap_profile;
+  delete data.pi_user_data;
+  if (!isNullOrEmptyString(instance.pi_user_data, true)) {
+    data.pi_user_data = `\${<<USER_DATA\n${instance.pi_user_data}\n  USER_DATA}`;
+  }
   return data;
 }